Refactor testutils lib, moved io code into submodule
